Fix small error in setting the output filename
The previous method of using strip() had potential of removing letters
from the beginning of a filename.  For example, the input file of
"p1020440_89.8616_o.jpg" would create an output file named
"1020440_89.8616_o.avif" as strip() would match and remove the
inital p (based on the search pattern "jpg").

Instead, let's just remove letters based on the length of the
file extension.